(#102) Add HdlLibrary object and its VHDL parsing
Add an HdlLibrary C++ object to represent a VHDL library declaration.
Add the parsing logic to create this object from a VHDL library clause.

Add a corresponding Python object and code to convert to and from the
Python AST.  Also reworked the Python around the VHDL 'use' clause
emission to eliminate "auto" library emission.  Since the library
object is now explicit, it should be created in the AST for it to
appear in the VHDL output.